Solve the hd0j1010 with a recursive version

+++ cpputest/interview1/interview1/main.cpp
@@ -9,7 +9,7 @@ void inverseCopy(char* src, char* dest)
 {
     assert(src != NULL && dest != NULL);
     
-    int len = strlen(src);
+    int len = (int)strlen(src);
     int i = len - 1;
     int k = 0;
